refactor(anki): refactor anki sm2 logic
