+++ src/services/spaced-repetition-system/super-memo-2/anki/constants/anki.constant.ts
@@ -1,0 +1,3 @@
+export const INVALID_RATING_MESSAGE = 'Invalid rating option';
+export const INVALID_STATUS_MESSAGE = 'Invalid card status';
+export const NULL_STEP_MESSAGE = 'Card step is NULL';
